requirements/base.txt
Django==4.2
MouseInfo==0.1.3
Pillow==10.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyRect==0.2.0
PyScreeze==1.0.1
PyYAML==6.0.2
aiosmtplib==1.1.7
anyio==3.7.1
asgiref==3.8.1
cache==1.0.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
djangorestframework==3.14.0
exceptiongroup==1.2.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.10
keyboard==0.13.5
messages==0.8.0
mouse==0.7.1
mss==9.0.2
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
plotly==5.24.1
psutil==6.1.0
psycopg2==2.9.7
pyperclip==1.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
pywin32==308
reportlab==4.2.5
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validus==0.3.0
